requirements.txt
nb-cli
nonebot2[fastapi]
nonebot-adapter-onebot
nonebot-plugin-apscheduler
aiohttp
pydantic
